test(tools): fold access-matrix pin into tool-access tests
